Consolidate hmac.c
- Use CommonCrypto on Darwin
- Reuse mbed HMAC on Windows/Linux/Android
- Delete unused pp_hmac_create
