+++ Sources/PartoutCore_C/include/crypto/hmac.h
@@ -6,8 +6,8 @@
 
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
-#include "portable/zd.h"
 
 #pragma clang assume_nonnull begin
 
@@ -21,7 +21,6 @@ typedef struct {
     size_t data_len;
 } pp_hmac_ctx;
 
-pp_zd *pp_hmac_create(void);
 size_t pp_hmac_do(pp_hmac_ctx *ctx);
 
 #pragma clang assume_nonnull end
