+++ Sources/PartoutCrypto/Native_C/src/android/hmac.c
