+++ Sources/PartoutCrypto/Native_C/src/linux/hmac.c
