+++ Sources/partout.cmake
@@ -402,11 +402,14 @@ set(PARTOUT_C_SOURCES
 ./PartoutCore_C/tun_windows.c
 ./PartoutCore_C/zd.c
 ./PartoutCrypto/Native_C/src/android/crypto.c
+./PartoutCrypto/Native_C/src/android/hmac.c
 ./PartoutCrypto/Native_C/src/darwin/crypto.c
 ./PartoutCrypto/Native_C/src/darwin/crypto_aead.c
 ./PartoutCrypto/Native_C/src/darwin/crypto_cbc.c
 ./PartoutCrypto/Native_C/src/darwin/crypto_ctr.c
+./PartoutCrypto/Native_C/src/darwin/hmac.c
 ./PartoutCrypto/Native_C/src/linux/crypto.c
+./PartoutCrypto/Native_C/src/linux/hmac.c
 ./PartoutCrypto/Native_C/src/mbed/keys.c
 ./PartoutCrypto/Native_C/src/mbed/tls.c
 ./PartoutCrypto/Native_C/src/windows/crypto.c
